Refactor MycoBank external link helper
- Make it return link, instead of url
- Change its name to reflect changed function
- Delete unneeded substitution of `%20` for spaces and change tests accordingly.  (Rails (or someone) is now escaping the spaces in the url.)
- Create file to test ExternalLinkHelper, and move tests there from ObjectLinkHelperTest
