+++ app/views/name/show_name.html.erb
@@ -70,7 +70,7 @@
             <%= content_tag(:div, link_to('EOL', @name.eol_url)) if @name.eol_url %>
             <%= content_tag(:div, link_to(:google_images.t, "http://images.google.com/images?q=#{@name.real_text_name}")) %>
             <%= content_tag(:div, link_to("MyCoPortal", mycoportal_url(@name))) %>
-            <%= content_tag(:div, link_to("MycoBank", mycobank_url(@name))) %>
+            <%= content_tag(:div, link_to_mycobank_search(@name)) %>
             <%= content_tag(:div, link_with_query(:show_name_distribution_map.t, action: 'map', id: @name.id)) %>
           </div>
         </div>
