+++ app/views/observer/_show_name_info.html.erb
@@ -6,8 +6,8 @@
                      action: :show_name, id: name.id)
     lines << link_to(name.display_name.t + " " + :show_name_on_mycoportal.t,
                      mycoportal_url(name))
-    lines << link_to(name.text_name.t + " " + :show_name_on_mycobank.t,
-                     mycobank_url(name))
+    lines << link_to_mycobank_search(name, name.text_name.t + " " +
+                                           :show_name_on_mycobank.t)
  end
 
   if name && !name.unknown? && !browser.bot?
